api/shared/config.py
import os
from pathlib import Path
from dotenv import load_dotenv

load_dotenv()

DATA_DIR = Path(os.getenv("CRITICAL_PATH_DATA_DIR", Path(__file__).parent.parent.parent / "data"))
OPENAI_MODEL = os.getenv("OPENAI_MODEL", "gpt-4o-mini")
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "")
MAX_HISTORY_MESSAGES = int(os.getenv("MAX_HISTORY_MESSAGES", "20"))
AZURE_STORAGE_CONNECTION_STRING = os.getenv("AZURE_STORAGE_CONNECTION_STRING", "")


if __name__ == "__main__":
    print("Data Dir: ",DATA_DIR)
    print("OpenAI Model: ", OPENAI_MODEL)
